Configurações de segurança

+++ src/main/java/com/finance/api/domain/dadosTransacoes/DadosCompletosTransacao.java
@@ -1,8 +1,9 @@
 package com.finance.api.domain.dadosTransacoes;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
-public record DadosCompletosTransacao(Long id,String descricao, Double valor, LocalDateTime data, Tipo tipo) {
+public record DadosCompletosTransacao(Long id, String descricao, BigDecimal valor, LocalDateTime data, Tipo tipo) {
 
     public DadosCompletosTransacao(Transacao transacao) {
         this(transacao.getId(),transacao.getDescricao(),transacao.getValor(),transacao.getData(),transacao.getTipo());
